Extracted super-class out of NonBlockingStatsDClient to enable multiple implementations.
